Moved scene description from main.cpp

+++ utilities/vec3.h
@@ -15,8 +15,8 @@ public:
     double z() const { return e[2]; }
 
     vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-    double operator[](int i) const { return e[i]; }
-    double& operator[](int i) { return e[i]; }
+    double operator[](size_t i) const { return e[i]; }
+    double& operator[](size_t i) { return e[i]; }
 
     vec3& operator+=(const vec3& v) {
         e[0] += v.e[0];
